+++ test/multinetwork.jl
@@ -203,6 +203,23 @@ TESTLOG = Memento.getlogger(PowerModels)
             )
         end
 
+        @testset "test sdp with constraint decomposition opf" begin
+            result = PowerModels.run_mn_opf(mn_data, SparseSDPWRMPowerModel, scs_solver)
+
+            @test result["termination_status"] == OPTIMAL
+            @test isapprox(result["objective"], 33321.9; atol = 1e0)
+            @test isapprox(
+                result["solution"]["nw"]["1"]["gen"]["2"]["pg"],
+                result["solution"]["nw"]["2"]["gen"]["2"]["pg"];
+                atol = 1e-3
+            )
+            @test isapprox(
+                result["solution"]["nw"]["1"]["gen"]["4"]["pg"],
+                result["solution"]["nw"]["2"]["gen"]["4"]["pg"];
+                atol = 1e-3
+            )
+        end
+
         @testset "test nfa opf" begin
             result = PowerModels.run_mn_opf(mn_data, NFAPowerModel, ipopt_solver)